refactor: 역할에 따른 클래스 분리
- 문자열을 분할
- 문자 숫자 타입 변경
- 숫자 합계
